+++ pytest_httpx/__init__.py
@@ -1,4 +1,2 @@
 from pytest_httpx.version import __version__
 from pytest_httpx._httpx_mock import httpx_mock, HTTPXMock
-
-# TODO Auto register fixture the way pytest-responses does