+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] - 2020-02-10
+### Added
+- match_headers parameter now allows to match on headers.
+- match_content parameter now allows to match on full body.
+
+### Changed
+- httpx.HTTPError is now raised instead of Exception in case a request cannot be matched.
+
 ## [0.0.4] - 2020-02-07
 ### Changed
 - url is not a mandatory parameter for response registration anymore.
@@ -42,7 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release, should be considered as unstable for now as design might change.
 
-[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/Colin-b/pytest_httpx/compare/v0.0.1...v0.0.2